requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
joblib 
Flask
Flask-SocketIO
python-socketio
eventlet 
paho-mqtt